\name{caRpools}
\alias{CRISPR-AnalyzeR-package}
\alias{CRISPR-AnalyzeR}
\docType{package}
\title{
CaRpools - CRISPR-AnalyzeR for pooled Screens
}
\description{
Analysis of pooled CRISPR screens based on mapped NGS readcount data or raw NGS FASTQ file.
}
\details{
\tabular{ll}{
Package: \tab caRpools\cr
Type: \tab Package\cr
Version: \tab 0.81\cr
Date: \tab 2015-07-31\cr
License: \tab GPL V3\cr
}

}
\author{
Jan Winter (German Cancer Research Center, DKFZ),
Florian Heigwer (German Cancer Research Center, DKFZ)

Maintainer: Jan Winter <jan.winter@dkfz-heidelberg.de>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\keyword{ CRISPR }
\keyword{ Analysis }

