\name{libFILE}
\docType{data}
\alias{libFILE}
\title{FASTA file containing als sgRNA target sequences and identifiers. USed for mapping and sgRNA table.}
\description{
 FASTA file containing als sgRNA target sequences and identifiers. USed for mapping and sgRNA table.
}
\usage{libFILE}
\format{data frame}

\keyword{datasets}